Split up light data mod
